+++ app/Http/Controllers/Setup/CategoryController.php
@@ -10,21 +10,21 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Carbon;
 use Illuminate\Validation\Rule;
 use Illuminate\Routing\Controllers\HasMiddleware;
-use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller implements HasMiddleware
 {
     public static function middleware(): array
     {
         return [
-            'auth',
-            new Middleware('role:Super-Admin|Admin'),
+            'auth'
         ];
     }
 
 
     public function CategoryIndex()
     {
+        Gate::authorize('category.list');
         $categories = Category::join('category_types', 'category_types.id', '=', 'categories.category_type_id')
             ->select('categories.*', 'category_types.name')
             ->get();
@@ -34,28 +34,29 @@ class CategoryController extends Controller implements HasMiddleware
 
     public function CategoryAdd()
     {
+        Gate::authorize('category.create');
         $category_type = CategoryType::all();
         return view('admin.setups.category.add', compact('category_type'));
     }
 
     public function CategoryStore(Request $request)
     {
+        Gate::authorize('category.create');
+
         $validatedData = $request->validate([
             'category_name' => 'required',
         ]);
 
         $data = new Category();
-        $data->category_name = $request->category_name;
+        $data->category_name = ucwords($request->category_name);
         $data->category_type_id = $request->category_type_name;
         $data->category_code = $request->category_code;
         /******************************************************/
-        $maxnumber = DB::table('categories')
-            ->where('category_type_id', $request->category_type_name)
-            ->select(DB::raw('MAX(sorting_order) as value'))
-            ->get();
+        $maxnumber = Category::where('category_type_id', $request->category_type_name)
+                    ->select(DB::raw('MAX(sorting_order) as value'))
+                    ->get();
 
         $maxValue = json_decode($maxnumber, true);
-
         $sortnumber = $maxValue[0]["value"] + 1;
         /*******************************************************/
         $data->sorting_order = $sortnumber;
@@ -72,6 +73,7 @@ class CategoryController extends Controller implements HasMiddleware
 
     public function CategoryEdit($id)
     {
+        Gate::authorize('category.updation');
         $editData = Category::find($id);
         $categoryTypes = CategoryType::all();
         return view('admin.setups.category.edit', compact('editData', 'categoryTypes'));
@@ -79,13 +81,14 @@ class CategoryController extends Controller implements HasMiddleware
 
     public function CategoryUpdate(Request $request, $id)
     {
+        Gate::authorize('category.updation');
 
         $validatedData = $request->validate([
             'category_name' => ['required', Rule::unique('categories')->ignore($id)],
         ]);
 
         $data = Category::find($id);
-        $data->category_name = $request->category_name;
+        $data->category_name = ucwords($request->category_name);
         $data->category_code = $request->category_code;
         $data->save();
 
@@ -100,6 +103,7 @@ class CategoryController extends Controller implements HasMiddleware
 
     public function CategoryDelete($id)
     {
+        Gate::authorize('category.delete');
 
         $catagory = Category::find($id);
         $catagory->delete();
